feat: organize all css according to new template, add snippets for future css, move wildcard joker under resource joker
