+++ docs/StyledComponents.md
@@ -1,0 +1,80 @@
+# Styled Components
+
+This project uses styled components. The following template is used to organize the styled components into somewhat of a logical pattern for maintainability.
+
+```css
+
+/** Layout */
+
+/** Box Model */
+
+/** Colour + Background */
+
+/** Typography */
+
+/** Visual Effects */
+
+/** Responsive Design */
+
+/** Interactivity */
+
+/** Micellaneous */
+
+```
+
+Here is an extended example of how the template should be populated:
+
+```css
+
+/** Layout - structure and positioning */
+    display;
+    position;
+    justify;
+    align;
+    z-index;
+
+/** Box Model - dimensions and spacing */
+    height;
+    width;
+    padding;
+    margin;
+    border;
+    box-sizing
+
+/** Colour + Background */
+    color;
+    background-color;
+    background-image;
+
+/** Typography - appearance of text */
+    font-family;
+    font-size;
+    font-weight;
+    line-height;
+    text-align;
+    text-decoration;
+    text-shadow;
+    letter-spacing;
+
+/** Visual Effects - visual enhancements and animations */
+    box-shadow;
+    opacity;
+    drop-shadow;
+    filter;
+    visiblity;
+    transition;
+    animation;
+
+/** Responsive design - media query specific rules */
+
+/** Interactivity - user interactions */
+    hover;
+    cursor;
+    focus;
+    active;
+    disabled;
+
+/** Micellaneous */
+    clip-path;
+    overflow;
+```
