+++ frontend/src/components/App/styles/StyledApp.tsx
@@ -1,10 +1,25 @@
 import { styled } from "styled-components"
 
 const StyledApp = styled.div`
+/** Layout */
     display: flex;
     flex-flow: column;
+
+/** Box Model */
     min-height: 100dvh;
-    height: 100%;
+
+/** Colour + Background */
+
+/** Typography */
+
+/** Visual Effects */
+
+/** Responsive Design */
+
+/** Interactivity */
+
+/** Micellaneous */
+
 `
 
 export default StyledApp;
