+++ frontend/src/components/Asset/StyledAsset.tsx
@@ -4,8 +4,23 @@ import { styled } from "styled-components"
  * Any icon, graphic, or image that represents a game piece
  */
 const StyledAsset = styled.img`
+/** Layout */
     position: absolute;
+
+/** Box Model */
     width: 100%;
+
+/** Colour + Background */
+
+/** Typography */
+
+/** Visual Effects */
+
+/** Responsive Design */
+
+/** Interactivity */
+
+/** Micellaneous */
 `
 
 export default StyledAsset
