+++ frontend/src/components/Buttons/BuildButton/styles/StyledBuildButton.tsx
@@ -2,10 +2,25 @@ import { styled } from "styled-components"
 import StyledBaseButton from "../../styles/StyledBaseButton"
 
 const StyledBuildButton = styled(StyledBaseButton)`
+/** Layout */
     display: flex;
     flex-direction: row;
     justify-content: center;
+
+/** Box Model */
     width: 23%;
+
+/** Colour + Background */
+
+/** Typography */
+
+/** Visual Effects */
+
+/** Responsive Design */
+
+/** Interactivity */
+
+/** Micellaneous */
 `
 
 export default StyledBuildButton
