+++ frontend/src/components/Buttons/BuildButton/styles/StyledBuildButtonIcon.tsx
@@ -1,19 +1,34 @@
 import { styled, css } from 'styled-components'
 
 interface StyledBuildButtonIconProps {
-    $width: number;
-    $opacity: number;
+  $width: number;
+  $opacity: number;
 }
 
 const StyledBuildButtonIcon = styled.img<StyledBuildButtonIconProps>`
-  width: ${(props) => `${props.$width}%`};
-  opacity: ${(props) =>
-        props.$opacity &&
-        css`
-      ${props.$opacity}%
-    `};
+/** Layout */
 
+/** Box Model */
+  width: ${(props) => `${props.$width}%`};
+
+/** Colour + Background */
+
+/** Typography */
+
+/** Visual Effects */
+  opacity: ${(props) =>
+    props.$opacity &&
+    css`
+      ${props.$opacity}%
+      `};
+      
   transition: opacity 250ms ease-out;
+
+/** Responsive Design */
+
+/** Interactivity */
+
+/** Micellaneous */
 `;
 
 export default StyledBuildButtonIcon;
