+++ frontend/src/components/Buttons/PlayAgainButton/styles/StyledPlayAgainButton.tsx
@@ -2,10 +2,26 @@ import { styled } from "styled-components";
 import StyledBaseButton from "../../styles/StyledBaseButton";
 
 const StyledPlayAgainButton = styled(StyledBaseButton)`
-  width: 40%;
+/** Layout */
   align-self: center;
+
+/** Box Model */
+  width: 40%;
   padding-right: 2%;
   padding-left: 2%;
+
+/** Colour + Background */
+
+/** Typography */
+
+/** Visual Effects */
+
+/** Responsive Design */
+
+/** Interactivity */
+
+/** Micellaneous */
+
 `;
 
 export default StyledPlayAgainButton;
