+++ frontend/src/components/Buttons/PlayAgainButton/styles/StyledPlayAgainIcon.tsx
@@ -1,7 +1,22 @@
 import { styled } from 'styled-components'
 
 const StyledPlayAgainIcon = styled.img`
+/** Layout */
+
+/** Box Model */
   width: 20%;
+
+/** Colour + Background */
+
+/** Typography */
+
+/** Visual Effects */
+
+/** Responsive Design */
+
+/** Interactivity */
+
+/** Micellaneous */
 `;
 
 export default StyledPlayAgainIcon;
