+++ frontend/src/components/Buttons/ResetGameButton/styles/StyledResetGameButton.tsx
@@ -2,11 +2,24 @@ import { styled } from "styled-components";
 import StyledBaseButton from "../../styles/StyledBaseButton";
 
 const StyledResetGameButton = styled(StyledBaseButton)`
+/** Box Model */
   width: 40%;
   padding-left: 2%;
   padding-right: 2%;
-  white-space: nowrap;
+
+/** Colour + Background */
+
+/** Typography */
   text-align: center;
+  white-space: nowrap;
+
+/** Visual Effects */
+
+/** Responsive Design */
+
+/** Interactivity */
+
+/** Micellaneous */
 `;
 
 export default StyledResetGameButton;
