+++ frontend/src/components/Buttons/RollButton/styles/StyledRollButton.tsx
@@ -2,11 +2,26 @@ import { styled } from "styled-components"
 import StyledBaseButton from "../../styles/StyledBaseButton";
 
 const StyledRollButton = styled(StyledBaseButton)`
+/** Layout */
     display: flex;
     flex-direction: row;
     justify-content: center;
     gap: 3%;
+    
+/** Box Model */
     width: 35%;
+
+/** Colour + Background */
+
+/** Typography */
+
+/** Visual Effects */
+
+/** Responsive Design */
+
+/** Interactivity */
+
+/** Micellaneous */
 `
 
 export default StyledRollButton;
