+++ frontend/src/components/Buttons/RollButton/styles/StyledRollButtonIcon.tsx
@@ -5,7 +5,16 @@ interface StyledRollButtonIconProps {
 }
 
 const StyledRollButtonIcon = styled.img<StyledRollButtonIconProps>`
+/** Layout */
+
+/** Box Model */
   width: 24%;
+
+/** Colour + Background */
+
+/** Typography */
+
+/** Visual Effects */
   opacity: ${(props) =>
     props.$opacity &&
     css`
@@ -13,6 +22,13 @@ const StyledRollButtonIcon = styled.img<StyledRollButtonIconProps>`
     `};
 
   transition: opacity 250ms ease-out;
+
+/** Responsive Design */
+
+/** Interactivity */
+
+/** Micellaneous */
+
 `;
 
 export default StyledRollButtonIcon;
