+++ frontend/src/components/Buttons/VolumeButton/styles/StyledVolumeButton.tsx
@@ -1,8 +1,23 @@
 import { styled } from 'styled-components'
 
 const StyledVolumeButton = styled.div`
+/** Layout */ 
+
+/** Box Model */ 
     width: 7%;
     margin-right: 2%;
+
+/** Colour + Background */ 
+
+/** Typography */ 
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
+
+/** Micellaneous */
 `
 
 export default StyledVolumeButton
