+++ frontend/src/components/Buttons/VolumeButton/styles/StyledVolumeIcon.tsx
@@ -1,8 +1,23 @@
 import { styled } from 'styled-components'
 
 const StyledVolumeIcon = styled.img`
+/** Layout */ 
+
+ /** Box Model */ 
     width: 100%;
+
+ /** Colour + Background */ 
+
+ /** Typography */ 
+
+ /** Visual Effects */ 
+
+ /** Responsive Design */ 
+
+ /** Interactivity */ 
     cursor: pointer;
+
+ /** Micellaneous */
 `
 
 export default StyledVolumeIcon
