+++ frontend/src/components/Buttons/styles/StyledBaseButton.tsx
@@ -1,15 +1,27 @@
 import { styled } from "styled-components"
 
 const StyledBaseButton = styled.button`
-    background-color: #e8d8b5;
+/** Layout */
+
+/** Box Model */
     border: none;
     border-radius: 10px;
     padding-left: 1%;
     padding-right: 1%;
     padding-top: 1%;
     padding-bottom: 1%;
+
+/** Colour + Background */
+    background-color: #e8d8b5;
+
+/** Typography */
+
+/** Visual Effects */
     transition-duration: 0.4s;
-    
+
+/** Responsive Design */
+
+/** Interactivity */
     &:hover {
         background-color: #dcc7a1;
         cursor: pointer;
@@ -19,6 +31,8 @@ const StyledBaseButton = styled.button`
         cursor: not-allowed;
         pointer-events: none;
     }
-    `
+
+/** Micellaneous */
+`
 
 export default StyledBaseButton
