+++ frontend/src/components/GameControlsContainer/styles/StyledDiceTray.tsx
@@ -1,14 +1,24 @@
 import { styled } from "styled-components"
 
 const StyledDiceTray = styled.div`
+/** Layout */ 
     display: flex;
     flex-direction: row;
     justify-self: center;
     justify-content: center; 
     align-items: center;
     gap: 3%;
+
+/** Box Model */ 
     margin-top: 2%;
-    
+
+/** Colour + Background */ 
+
+/** Typography */ 
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
     @media ${(props) => props.theme.breakpoints.xs} {
         width: 90%;
     }
@@ -32,6 +42,10 @@ const StyledDiceTray = styled.div`
     @media ${(props) => props.theme.breakpoints.xxl} {
         width: 60%;
     }
+
+/** Interactivity */ 
+
+/** Micellaneous */    
 `
 
 export default StyledDiceTray;
