+++ frontend/src/components/GameControlsContainer/styles/StyledGameControlsContainer.tsx
@@ -1,11 +1,21 @@
 import { styled } from "styled-components"
 
 const StyledGameControlsContainer = styled.div`
-    padding: 1%;
+/** Layout */ 
     display: flex;
     flex-direction: column;
     align-items: center;
-    
+
+/** Box Model */ 
+    padding: 1%;
+
+/** Colour + Background */ 
+
+/** Typography */ 
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
     @media ${(props) => props.theme.breakpoints.xs} {
         width: 100%;
     }
@@ -29,6 +39,10 @@ const StyledGameControlsContainer = styled.div`
     @media ${(props) => props.theme.breakpoints.xxl} {
 
     }
+
+/** Interactivity */ 
+
+/** Micellaneous */   
 `
 
 export default StyledGameControlsContainer
