+++ frontend/src/components/Hexagon/styles/StyledHexagon.tsx
@@ -6,10 +6,25 @@ interface StyledHexagonProps {
 }
 
 const StyledHexagon = styled.div<StyledHexagonProps>`
+/** Layout */ 
     position: absolute;
     top: ${props => `${props.$top}%`};
     left: ${props => `${props.$left}%`};
+
+/** Box Model */ 
     width: 37.5%;
+
+/** Colour + Background */ 
+
+/** Typography */ 
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
+
+/** Micellaneous */
 `
 
 export default StyledHexagon
