+++ frontend/src/components/Layout/styles/StyledLayout.tsx
@@ -1,12 +1,28 @@
 import { styled } from "styled-components"
 
 const StyledLayout = styled.div`
+/** Layout */ 
     display: flex;
     flex-direction: column;
     flex-grow: 1;
     align-items: center;
     justify-content: flex-start;
+
+/** Box Model */ 
+
+/** Colour + Background */ 
     background-color: #6A8CA6;
+
+/** Typography */ 
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
+
+/** Micellaneous */
+
 `
 
 export default StyledLayout
