+++ frontend/src/components/Layout/styles/StyledVerticallyCenteredContainer.tsx
@@ -4,9 +4,24 @@ import { styled } from 'styled-components'
  * Used to center main content of the app vertically
  */
 const StyledVerticallyCenteredContainer = styled.div`
+/** Layout */ 
+
+/** Box Model */ 
     margin-top: auto;
     margin-bottom: auto;
     width: 100%;
+
+/** Colour + Background */ 
+
+/** Typography */ 
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
+
+/** Micellaneous */
 `
 
 export default StyledVerticallyCenteredContainer
