+++ frontend/src/components/Modals/GameOverModal/styles/StyledGameOverModalBody.tsx
@@ -2,7 +2,22 @@ import { ModalBody } from 'react-bootstrap'
 import { styled } from 'styled-components'
 
 const StyledGameOverModalBody = styled(ModalBody)`
+/** Layout */ 
+
+/** Box Model */ 
+
+/** Colour + Background */ 
+
+/** Typography */ 
     text-align: center;
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
+
+/** Micellaneous */
 `
 
 export default StyledGameOverModalBody
