+++ frontend/src/components/Modals/HowToPlayModal/styles/StyledHowToPlayAsset.tsx
@@ -1,12 +1,21 @@
 import { styled } from "styled-components"
 
 const StyledHowToPlayAsset = styled.img`
+/** Layout */ 
     display: block;
+
+/** Box Model */ 
     margin-left: auto;
     margin-right: auto;
     margin-bottom: 2%;
 
+/** Colour + Background */ 
 
+/** Typography */ 
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
     @media ${(props) => props.theme.breakpoints.xs} {
         width: 100%;
     }
@@ -24,12 +33,16 @@ const StyledHowToPlayAsset = styled.img`
     }
 
     @media ${(props) => props.theme.breakpoints.xl} {
-    
+
     }
 
     @media ${(props) => props.theme.breakpoints.xxl} {
-  
+
     }
+
+/** Interactivity */ 
+
+/** Micellaneous */
 `
 
 export default StyledHowToPlayAsset
