+++ frontend/src/components/Modals/HowToPlayModal/styles/StyledHowToPlayModalBody.tsx
@@ -2,15 +2,30 @@ import ModalBody from 'react-bootstrap/ModalBody'
 import { styled } from 'styled-components'
 
 const StyledHowToPlayModalBody = styled(ModalBody)`
-    > h1 {
-        font-size: 24px;
-        font-weight: bold;
-    }
+/** Layout */ 
 
-    > h2 {
-        font-size: 20px;
-        font-weight: bold;
-    }
+/** Box Model */ 
+
+/** Colour + Background */ 
+
+/** Typography */ 
+> h1 {
+    font-size: 24px;
+    font-weight: bold;
+}
+
+> h2 {
+    font-size: 20px;
+    font-weight: bold;
+}
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
+
+/** Micellaneous */
 `
 
 export default StyledHowToPlayModalBody
