+++ frontend/src/components/Modals/SettingsModal/styles/StyledSettingsModalBody.tsx
@@ -2,9 +2,24 @@ import ModalBody from "react-bootstrap/ModalBody";
 import { styled } from "styled-components";
 
 const StyledSettingsModalBody = styled(ModalBody)`
+/** Layout */ 
   display: flex;
   flex-direction: column;
   align-items: center;
+
+/** Box Model */ 
+
+/** Colour + Background */ 
+
+/** Typography */ 
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
+
+/** Micellaneous */
 `;
 
 export default StyledSettingsModalBody;
