+++ frontend/src/components/Modals/SettingsModal/styles/StyledVersion.tsx
@@ -1,7 +1,22 @@
 import { styled } from 'styled-components'
 
 const StyledVersion = styled.p`
+/** Layout */ 
+
+/** Box Model */ 
+
+/** Colour + Background */ 
+
+/** Typography */ 
     font-size: 12px;
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
+
+/** Micellaneous */
 `
 
 export default StyledVersion
