+++ frontend/src/components/Modals/styles/StyledModalFooter.tsx
@@ -2,9 +2,24 @@ import { ModalFooter } from 'react-bootstrap'
 import { styled } from 'styled-components'
 
 const StyledModalFooter = styled(ModalFooter)`
+/** Layout */ 
     display: flex;
     align-items: center;
     justify-content: center;
+
+/** Box Model */ 
+
+/** Colour + Background */ 
+
+/** Typography */ 
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
+
+/** Micellaneous */
 `
 
 export default StyledModalFooter
