+++ frontend/src/components/Modals/styles/StyledModalHeader.tsx
@@ -2,10 +2,25 @@ import { ModalHeader } from 'react-bootstrap'
 import { styled } from 'styled-components'
 
 const StyledModalHeader = styled(ModalHeader)`
-    font-size: 24px;
-    font-weight: bold;
+/** Layout */ 
     display: flex;
     justify-content: center;
+
+/** Box Model */ 
+
+/** Colour + Background */ 
+
+/** Typography */ 
+    font-size: 24px;
+    font-weight: bold;
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
+
+/** Micellaneous */
 `
 
 export default StyledModalHeader
