+++ frontend/src/components/Navbar/NavbarButton/styles/StyledNavbarButton.tsx
@@ -1,14 +1,29 @@
 import { styled } from "styled-components";
 
 const StyledNavbarButton = styled.div`
+/** Layout */ 
+
+/** Box Model */ 
   width: 100%;
 
+/** Colour + Background */ 
+
+/** Typography */ 
+
+/** Visual Effects */ 
+  transition: font-weight 250ms ease-out;
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
   &:hover {
     cursor: pointer;
     font-weight: bold;
   }
 
-  transition: font-weight 250ms ease-out;
+/** Micellaneous */
+
+
 `;
 
 export default StyledNavbarButton;
