+++ frontend/src/components/Navbar/styles/StyledNavbar.tsx
@@ -2,11 +2,26 @@ import Navbar from "react-bootstrap/Navbar";
 import { styled } from "styled-components";
 
 const StyledNavbar = styled(Navbar)`
+/** Layout */ 
   display: flex;
-  background-color: #e8d8b5;
+
+/** Box Model */ 
   width: 100%;
   padding-left: 2%;
   padding-right: 2%;
+
+/** Colour + Background */ 
+  background-color: #e8d8b5;
+
+/** Typography */ 
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
+
+/** Micellaneous */
 `;
 
 export default StyledNavbar;
