+++ frontend/src/components/Navbar/styles/StyledNavbarBrand.tsx
@@ -2,11 +2,26 @@ import { NavbarBrand } from "react-bootstrap";
 import { styled } from "styled-components";
 
 const StyledNavbarBrand = styled(NavbarBrand)`
+/** Layout */ 
   display: flex;
   align-items: center;
-  font-size: 20px;
   gap: 1%;
+
+/** Box Model */ 
+
+/** Colour + Background */ 
+
+/** Typography */ 
+  font-size: 20px;
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
   user-select: none;
+
+/** Micellaneous */
 `;
 
 export default StyledNavbarBrand;
