+++ frontend/src/components/Navbar/styles/StyledNavbarBrandIcon.tsx
@@ -1,8 +1,23 @@
 import { styled } from "styled-components"
 
 const StyledNavbarBrandIcon = styled.img`
+/** Layout */ 
+
+/** Box Model */ 
     width: 12%;
     margin-right: 2%;
+
+/** Colour + Background */ 
+
+/** Typography */ 
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
+
+/** Micellaneous */
 `
 
 export default StyledNavbarBrandIcon
