+++ frontend/src/components/Navbar/styles/StyledNavbarCollapse.tsx
@@ -2,9 +2,19 @@ import NavbarCollapse from 'react-bootstrap/NavbarCollapse'
 import { styled } from 'styled-components'
 
 const StyledNavbarCollapse = styled(NavbarCollapse)`
+/** Layout */ 
     display: flex;
     justify-content: right;
 
+/** Box Model */ 
+
+/** Colour + Background */ 
+
+/** Typography */ 
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
     @media ${(props) => props.theme.breakpoints.xs} {
         flex-direction: column;
     }
@@ -14,20 +24,25 @@ const StyledNavbarCollapse = styled(NavbarCollapse)`
     }
 
     @media ${(props) => props.theme.breakpoints.m} {
-    
+
     }
 
     @media ${(props) => props.theme.breakpoints.l} {
-      
+    
     }
 
     @media ${(props) => props.theme.breakpoints.xl} {
-     
+    
     }
 
     @media ${(props) => props.theme.breakpoints.xxl} {
-      
+    
     }
+
+/** Interactivity */ 
+
+/** Micellaneous */
+
 `
 
 export default StyledNavbarCollapse
