+++ frontend/src/components/Navbar/styles/StyledNavbarText.tsx
@@ -2,10 +2,19 @@ import NavbarText from 'react-bootstrap/NavbarText'
 import { styled } from 'styled-components'
 
 const StyledNavbarText = styled(NavbarText)`
-    user-select: none;
+/** Layout */ 
+
+/** Box Model */ 
+
+/** Colour + Background */ 
+
+/** Typography */ 
     white-space: nowrap;
 
-     @media ${(props) => props.theme.breakpoints.xs} {
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+    @media ${(props) => props.theme.breakpoints.xs} {
         margin-left: 0%;
     }
 
@@ -14,20 +23,27 @@ const StyledNavbarText = styled(NavbarText)`
     }
 
     @media ${(props) => props.theme.breakpoints.m} {
-    
+
     }
 
     @media ${(props) => props.theme.breakpoints.l} {
-      
+    
     }
 
     @media ${(props) => props.theme.breakpoints.xl} {
-     
+    
     }
 
     @media ${(props) => props.theme.breakpoints.xxl} {
-      
+    
     }
+
+/** Interactivity */ 
+    user-select: none;
+
+/** Micellaneous */
+
+
 `
 
 export default StyledNavbarText
