+++ frontend/src/components/Popups/WildcardTradingPopup/styles/StyledWildcardTradingIcon.tsx
@@ -1,12 +1,27 @@
 import { styled } from "styled-components"
 
 const StyledTradingIcon = styled.img`
+/** Layout */ 
+
+/** Box Model */ 
     width: 20%;
 
+/** Colour + Background */ 
+
+/** Typography */ 
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
     &:hover{
         cursor: pointer;
         transform: scale(1.1);
     }
+
+/** Micellaneous */
+
 `
 
 export default StyledTradingIcon
