+++ frontend/src/components/Popups/styles/StyledPopupArrow.tsx
@@ -7,11 +7,26 @@ interface StyledPopupArrowProps {
 }
 
 const StyledPopupArrow = styled.img<StyledPopupArrowProps>`
+/** Layout */ 
     position: absolute;
     top: ${props => `${props.$top}%`};
     left: ${props => `${props.$left}%`};
     width: ${props => `${props.$width}%`};
     z-index: 3;
+
+/** Box Model */ 
+
+/** Colour + Background */ 
+
+/** Typography */ 
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
+
+/** Micellaneous */
 `
 
 export default StyledPopupArrow
