+++ frontend/src/components/ResourceDice/styles/StyledResourceDice.tsx
@@ -5,15 +5,29 @@ interface StyledResourceDiceProps {
 }
 
 const StyledResourceDice = styled.div<StyledResourceDiceProps>`
+/** Layout */ 
     position: relative;
+
+/** Box Model */ 
     width: 100%;
-    
+
+/** Colour + Background */ 
+
+/** Typography */ 
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
     &:hover {
         cursor: ${props => props.$pointer &&
         css`
             pointer
         `};
     }
+
+/** Micellaneous */
 `
 
 export default StyledResourceDice
