+++ frontend/src/components/ResourceDice/styles/StyledResourceDiceFace.tsx
@@ -5,22 +5,30 @@ import pulse from "../../../animations/pulse"
 interface StyledResourceDiceFaceProps {
     $wobble: boolean
     $wobbleDurationMilliseconds: number
-    $grayscale: boolean
+    $isSpent: boolean
     $pulse: boolean
 }
 
 const StyledResourceDiceFace = styled.img<StyledResourceDiceFaceProps>`
+/** Layout */ 
+
+/** Box Model */ 
     width: 100%;
     border-radius: 10%;
 
+/** Colour + Background */ 
+
+/** Typography */ 
+
+/** Visual Effects */ 
     filter: ${props =>
-        props.$grayscale &&
+        props.$isSpent &&
         css`
             grayscale(100%)
         `};
         
     opacity: ${props =>
-        props.$grayscale &&
+        props.$isSpent &&
         css`
             30%
         `
@@ -29,7 +37,7 @@ const StyledResourceDiceFace = styled.img<StyledResourceDiceFaceProps>`
     animation: ${props =>
         (props.$wobble &&
             css`
-             ${wobble} ${props.$wobbleDurationMilliseconds}ms linear 
+            ${wobble} ${props.$wobbleDurationMilliseconds}ms linear 
         `)
         || (props.$pulse &&
             css`${pulse(1.05)} 1s infinite`
@@ -37,6 +45,13 @@ const StyledResourceDiceFace = styled.img<StyledResourceDiceFaceProps>`
     };
 
     transition: opacity 250ms ease-out, filter 250ms ease-out;
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
+
+/** Micellaneous */
+
 `
 
 export default StyledResourceDiceFace
