+++ frontend/src/components/ResourceJoker/WildcardJoker.tsx
@@ -4,7 +4,7 @@ import { useAppDispatch, useAppSelector } from "../../store/hooks"
 import { selectIsGamePhaseBuilding } from "../../store/slices/gameSlice/gameSlice"
 import { selectIsKnightBuilt } from "../../store/slices/knightSlice/knightSlice"
 import { selectIsResourceJokerSpent } from "../../store/slices/resourceJokerSlice/resourceJokerSlice"
-import StyledResourceJoker from "../ResourceJoker/styles/StyledResourceJoker"
+import StyledResourceJoker from "./styles/StyledResourceJoker"
 import StyledAsset from "../Asset/StyledAsset"
 import { useState } from "react"
 import { ResourceType } from "../../constants/resources"
