+++ frontend/src/components/Scoreboard/styles/StyledEqualsIcon.tsx
@@ -1,9 +1,24 @@
 import { styled } from "styled-components"
 
 const StyledEqualsIcon = styled.div`
+/** Layout */ 
+
+/** Box Model */ 
     width: 100%;
+
+/** Colour + Background */ 
+
+/** Typography */ 
     font-size: 20px;
     text-align: center;
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
+
+/** Micellaneous */
 `
 
 export default StyledEqualsIcon;
