+++ frontend/src/components/Scoreboard/styles/StyledScoreboard.tsx
@@ -1,12 +1,22 @@
 import { styled } from "styled-components"
 
 const StyledScoreboard = styled.div`
+/** Layout */ 
     display: grid;
     justify-items: stretch;
+
+/** Box Model */ 
     margin-left: auto;
     margin-right: auto;
     padding: 1%;
-    
+
+/** Colour + Background */ 
+
+/** Typography */ 
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
     @media ${(props) => props.theme.breakpoints.xs} {
         grid-template-columns: repeat(9, 1fr);
         grid-template-rows: repeat(2, 1fr);
@@ -36,6 +46,11 @@ const StyledScoreboard = styled.div`
         width: 60%;
         margin-bottom: 1%;
     }
+
+/** Interactivity */ 
+
+/** Micellaneous */
+    
 `
 
 export default StyledScoreboard
