+++ frontend/src/components/Scoreboard/styles/StyledTotalScore.tsx
@@ -1,12 +1,22 @@
 import { styled } from "styled-components"
 
 const StyledTotalScore = styled.div`
-    width: 100%;
+/** Layout */ 
     display: flex;
     justify-self: stretch;
     align-self: stretch;
+
+/** Box Model */ 
+    width: 100%;
     padding-left: 1.5%;
 
+/** Colour + Background */ 
+
+/** Typography */ 
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
     @media ${(props) => props.theme.breakpoints.xs} {
         grid-column: span 3;
     }
@@ -30,6 +40,11 @@ const StyledTotalScore = styled.div`
     @media ${(props) => props.theme.breakpoints.xxl} {
 
     }
+
+/** Interactivity */ 
+
+/** Micellaneous */
+
 `
 
 export default StyledTotalScore
