+++ frontend/src/components/Settings/styles/StyledSettings.tsx
@@ -1,9 +1,24 @@
 import { styled } from 'styled-components'
 
 const StyledSettings = styled.div`
+/** Layout */ 
     display: flex;
     flex-direction: column;
     justify-items: center;
+
+/** Box Model */ 
     width: 90%;
+
+/** Colour + Background */ 
+
+/** Typography */ 
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
+
+/** Micellaneous */
 `
 export default StyledSettings
