+++ frontend/src/components/Settings/styles/StyledSettingsLabel.tsx
@@ -1,8 +1,23 @@
 import { styled } from 'styled-components'
 
 const StyledSettingsLabel = styled.div`
+/** Layout */ 
+
+/** Box Model */ 
     margin-right: 2%;
+
+/** Colour + Background */ 
+
+/** Typography */ 
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
     user-select: none;
+
+/** Micellaneous */
 `
 
 export default StyledSettingsLabel;
