+++ frontend/src/components/Settings/styles/StyledSettingsRow.tsx
@@ -1,10 +1,25 @@
 import { styled } from 'styled-components'
 
 const StyledSettingsRow = styled.div`
+/** Layout */ 
     display: flex;
     align-items: center;
+
+/** Box Model */ 
     width: 100%;
     margin-bottom: 5%;
+
+/** Colour + Background */ 
+
+/** Typography */ 
+
+/** Visual Effects */ 
+
+/** Responsive Design */ 
+
+/** Interactivity */ 
+
+/** Micellaneous */
 `
 
 export default StyledSettingsRow
